Fix app; add build scripts
First, the app was broken. Just a black screen after pressing one of the
home page buttons. This fixes that.

Also adds some build scripts. They aren't working, but they work. It
throws some sass error, so i think the sass is not being built. But I
only needed to fix the js for this fix, so consider that can kicked down
the road. Sorry, future me.

The script is called build:local b/c i can't have heroku be building it
and failing bc of sass