Make the biggest update Asteroids has ever seen!
The original goal of this project was to be server free. I've always had
a love for the kind of recursive architecture that follows from truly
serverless (like, not "serverless", but literally without server, all
client side) applications. This is a game and there's no server, however
there is a host. It uses the technology of @browser-network, which I
created earlier this year, to finally achieve that. The impetus behind
doing it was that heroku, where the server version of this was hosted,
said no more free hosting. So instead of paying the ~$90 a year to host
it, I took enough time to remake it so that I could have hosted it
forever in opportunity cost. This seems worth it to me because it's my
project! Also this was a great test bed for @browser-network, which
received lots of primarily UX upgrades based on this work.

As it stands, this presumes switchboard.aaronik.com to be a
@browser-network switchboard.