Create a little widget for network status
This allows removing it from the game readout. This is important because
that was semantically incorrect - each time the game restarted, the
readout was reset, even though that event was not when there was a
change in the network status. Now the network thing lives for page
length, the same as the network status itself.

Also, the plan is in the future this NetworkInfoDropdown will be in its
own repo and can be shared across multiple apps.